Shuffle test fixtures
