+++ pensieve/tests/test_analysis.py
@@ -2,34 +2,9 @@ import datetime as dt
 from datetime import timedelta
 import json
 import pytz
-import pytest
 
 from pensieve.analysis import Analysis
 from pensieve.experimenter import Experiment
-
-
-@pytest.fixture
-def experiments():
-    return [
-        Experiment(
-            slug="test_slug",
-            type="pref",
-            start_date=dt.datetime(2019, 12, 1, tzinfo=pytz.utc),
-            end_date=dt.datetime(2020, 3, 1, tzinfo=pytz.utc),
-            proposed_enrollment=7,
-            variants=[],
-            normandy_slug="normandy-test-slug",
-        ),
-        Experiment(
-            slug="test_slug",
-            type="addon",
-            start_date=dt.datetime(2019, 12, 1, tzinfo=pytz.utc),
-            end_date=dt.datetime(2020, 3, 1, tzinfo=pytz.utc),
-            proposed_enrollment=0,
-            variants=[],
-            normandy_slug=None,
-        ),
-    ]
 
 
 def test_get_timelimits_if_ready(experiments):
